backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
asyncpg
psycopg2-binary
python-dotenv
pytest
pytest-asyncio
httpx
aiosqlite
email-validator
greenlet
